backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.1.0

langchain-openai>=0.1.10
langchain-community>=0.2.10
langchain-core>=0.2.10
langchain-text-splitters>=0.2.0
openai>=1.35.1
chromadb>=0.5.0

python-dotenv>=1.0.1
numpy>=1.26.4